feat: bestIds 자료형 수정

+++ backend/src/main/java/com/cocktailmasters/backend/vote/controller/dto/opinion/FindOpinionsResponse.java
@@ -11,6 +11,6 @@ import java.util.List;
 public class FindOpinionsResponse {
 
     List<OpinionDto> opinions;
-    List<Integer> bestIds;
+    List<Long> bestIds;
     int totalCount;
 }
